Move playlists logic to models.py
